backend/requirements.txt
# Core
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0

# Database
pandas==2.1.3
numpy==1.26.2
sqlalchemy==2.0.23
sqlite3

# ML/Recommendations
scikit-learn==1.3.2
scipy==1.11.4

# API Integration
httpx==0.25.2
python-multipart==0.0.6

# Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4